+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/ParBuilder.scala
@@ -9,8 +9,8 @@ import coop.rchain.models.Par
 import coop.rchain.models.rholang.implicits.VectorPar
 import coop.rchain.models.rholang.sorter.Sortable
 import coop.rchain.rholang.interpreter.errors._
-import coop.rchain.rholang.parser.rholang_mercury.Absyn.Proc
-import coop.rchain.rholang.parser.rholang_mercury.{parser, Yylex}
+import coop.rchain.rholang.ast.rholang_mercury.Absyn.Proc
+import coop.rchain.rholang.ast.rholang_mercury.{parser, Yylex}
 
 trait ParBuilder[F[_]] {
   def buildNormalizedTerm(source: String, normalizerEnv: Map[String, Par]): F[Par]
