+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalize.scala
@@ -9,7 +9,7 @@ import coop.rchain.models.Var.VarInstance._
 import coop.rchain.models._
 import coop.rchain.models.rholang.implicits._
 import coop.rchain.rholang.interpreter.errors._
-import coop.rchain.rholang.parser.rholang_mercury.Absyn.{
+import coop.rchain.rholang.ast.rholang_mercury.Absyn.{
   Bundle => AbsynBundle,
   Ground => AbsynGround,
   KeyValuePair => AbsynKeyValuePair,
