+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/ParBuilderUtil.scala
@@ -5,7 +5,7 @@ import java.io.{Reader, StringReader}
 import cats.effect.Sync
 import coop.rchain.models.Par
 import coop.rchain.rholang.interpreter.compiler.ParBuilder
-import coop.rchain.rholang.parser.rholang_mercury.Absyn.Proc
+import coop.rchain.rholang.ast.rholang_mercury.Absyn.Proc
 import monix.eval.Coeval
 
 object ParBuilderUtil {
